Add puzzle solving infrastructure and some solvers
Add code solving some of the puzzles of Eric Wastl's Advent of Code.
The puzzle solvers can be run from a shell script, `run.sh`, or from a
Python script, `run.py`.  Solvers are added for days 1 and 2 of Advent
of Code 2017 as well as days 1 to 11 of Advent of Code 2019.

The primary aim is to make the code readable, and the secondary aim is
to make it fast.  Some functions and other tools not specific to a
particular puzzle are grouped in separate modules, especially when doing
so improves readability and when these tools are reusable.

Some of the code added in this commit is older, with parts of it dating
back to 2017.

This project is published under the MIT License.